Actualizar

+++ SAGA.API/Controllers/Ventas/Requisiciones/RequisicionesController.cs
@@ -631,23 +631,13 @@ namespace SAGA.API.Controllers
             try
             {
                 var UnidadNegocio = db.Usuarios.Where(u => u.Id.Equals(propietario)).Select(u => u.Sucursal.UnidadNegocioId).FirstOrDefault();
-
-                if (UnidadNegocio != 3 )
-                {
-                    var Sucursales = db.OficinasReclutamiento.Where(s => s.UnidadNegocioId.Equals(UnidadNegocio)).Select(s => s.Id).ToList();
-                    //var Usuarios = db.Usuarios.Where(u => Sucursales.Contains(u.SucursalId)).Select(u => u.Id).ToList();
-
                     var requisicion = db.Requisiciones
                         .Where(e => e.Activo.Equals(true) && e.TipoReclutamientoId.Equals(tipo) && (e.EstatusId.Equals(43) || e.EstatusId.Equals(44)) )
-                        //.Where(e => Usuarios.Contains(e.PropietarioId))
                         .Select(e => new
                         {
                             Id = e.Id,
+                            EstadoId = e.Direccion.EstadoId,
                             VBtra = e.VBtra,
-                            //TipoReclutamiento = e.TipoReclutamiento.tipoReclutamiento,
-                            //tipoReclutamientoId = e.TipoReclutamientoId,
-                            //ClaseReclutamiento = e.ClaseReclutamiento.clasesReclutamiento,
-                            //ClaseReclutamientoId = e.ClaseReclutamientoId,
                             SueldoMinimo = e.SueldoMinimo,
                             SueldoMaximo = e.SueldoMaximo,
                             fch_Creacion = e.fch_Creacion,
@@ -666,99 +656,12 @@ namespace SAGA.API.Controllers
                             Vacantes = e.horariosRequi.Count() > 0 ? e.horariosRequi.Sum(h => h.numeroVacantes) : 0,
                             Folio = e.Folio,
                             Porcentaje = db.FacturacionPuro.Where(x => x.RequisicionId.Equals(e.Id)).Select(p => p.Porcentaje).FirstOrDefault(),
-                            //DiasEnvio = e.DiasEnvio,
-                            //Confidencial = e.Confidencial,
-                            //Postulados = db.Postulaciones.Where(p => p.RequisicionId.Equals(e.Id) & p.StatusId.Equals(1)).Count(),
-                            //PostuladosN = db.Postulaciones.Where(p => p.RequisicionId.Equals(e.Id) & p.StatusId.Equals(1)).Select(p => new
-                            //{
-                            //    p.CandidatoId,
-                            //    p.Candidato.Nombre,
-                            //    p.Candidato.ApellidoPaterno,
-                            //    p.Candidato.ApellidoMaterno,
-                            //    p.Candidato.CURP,
-                            //    email = p.Candidato.emails.Select(m => m.email).FirstOrDefault(),
-                            //    p.StatusId
-                            //}),
-                            //EnProceso = db.ProcesoCandidatos.Where(p => p.RequisicionId.Equals(e.Id) && p.EstatusId != 27 && p.EstatusId != 40 && p.EstatusId != 28 && p.EstatusId != 42).Count(),
-                            //EnProcesoN = db.ProcesoCandidatos.Where(p => p.RequisicionId.Equals(e.Id) && p.EstatusId != 24 && p.EstatusId != 27 && p.EstatusId != 40 && p.EstatusId != 28 && p.EstatusId != 42).Select(d => new
-                            //{
-                            //    candidatoId = d.CandidatoId,
-                            //    nombre = db.Candidatos.Where(x => x.Id.Equals(d.CandidatoId)).Select(cc => cc.Nombre + " " + cc.ApellidoPaterno + " " + cc.ApellidoMaterno).FirstOrDefault(),
-                            //    email = db.Emails.Where(x => x.EntidadId.Equals(d.CandidatoId)).Select(m => m.email).FirstOrDefault(),
-                            //    estatusId = d.EstatusId
-                            //}),
-                            //Contratados = db.ProcesoCandidatos.Where(p => p.RequisicionId.Equals(e.Id) && p.EstatusId.Equals(24)).Count(),
                             Propietario = db.Usuarios.Where(x => x.Id.Equals(e.PropietarioId)).Select(P => P.Nombre + " " + P.ApellidoPaterno + " " + P.ApellidoMaterno).FirstOrDefault(),
-                            //reclutadores = db.AsignacionRequis.Where(x => x.RequisicionId.Equals(e.Id) && !x.GrpUsrId.Equals(e.AprobadorId)).Select(a => new
-                            //{
-                            //    reclutador = db.Usuarios.Where(x => x.Id.Equals(a.GrpUsrId)).Select(r => r.Nombre + " " + r.ApellidoPaterno + " " + r.ApellidoMaterno).FirstOrDefault()
-                            //}).Distinct().ToList(),
-                            //ComentarioReclutador = db.ComentariosVacantes.Where(x => x.RequisicionId.Equals(e.Id)).Select(c => c.fch_Creacion + " - " + c.UsuarioAlta + " - " + (c.Motivo.Id == 7 ? "" : c.Motivo.Descripcion + " - ") + c.Comentario).ToList()
+                            ComentarioReclutador = db.ComentariosVacantes.Where(x => x.RequisicionId.Equals(e.Id)).Select(c => c.fch_Creacion + " - " + c.UsuarioAlta + " - " + (c.Motivo.Id == 7 ? "" : c.Motivo.Descripcion + " - ") + c.Comentario).ToList()
                         }).OrderBy(x => x.EstatusOrden).ThenByDescending(x => x.Folio).ToList();
 
                     return Ok(requisicion);
-                }else
-                {
-                    var requisicionMTY = db.Requisiciones
-                       .Where(e => e.Activo.Equals(true) && e.TipoReclutamientoId.Equals(tipo) && (e.EstatusId.Equals(43) || e.EstatusId.Equals(44)))
-                       .Select(e => new
-                       {
-                           Id = e.Id,
-                           VBtra = e.VBtra,
-                           //TipoReclutamiento = e.TipoReclutamiento.tipoReclutamiento,
-                           //tipoReclutamientoId = e.TipoReclutamientoId,
-                           //ClaseReclutamiento = e.ClaseReclutamiento.clasesReclutamiento,
-                           //ClaseReclutamientoId = e.ClaseReclutamientoId,
-                           SueldoMinimo = e.SueldoMinimo,
-                           SueldoMaximo = e.SueldoMaximo,
-                           fch_Creacion = e.fch_Creacion,
-                           fch_Modificacion = e.fch_Modificacion,
-                           fch_Cumplimiento = e.fch_Cumplimiento,
-                           Estatus = e.Estatus.Descripcion,
-                           EstatusId = e.EstatusId,
-                           EstatusOrden = e.Estatus.Orden,
-                           Prioridad = e.Prioridad.Descripcion,
-                           PrioridadId = e.PrioridadId,
-                           Cliente = e.Cliente.Nombrecomercial,
-                           razon = e.Cliente.RazonSocial,
-                           factura = e.Cliente.RazonSocial,
-                           GiroEmpresa = e.Cliente.GiroEmpresas.giroEmpresa,
-                           ActividadEmpresa = e.Cliente.ActividadEmpresas.actividadEmpresa,
-                           Vacantes = e.horariosRequi.Count() > 0 ? e.horariosRequi.Sum(h => h.numeroVacantes) : 0,
-                           Folio = e.Folio,
-                           Porcentaje = db.FacturacionPuro.Where(x => x.RequisicionId.Equals(e.Id)).Select(p => p.Porcentaje).FirstOrDefault(),
-                           //DiasEnvio = e.DiasEnvio,
-                           //Confidencial = e.Confidencial,
-                           //Postulados = db.Postulaciones.Where(p => p.RequisicionId.Equals(e.Id) & p.StatusId.Equals(1)).Count(),
-                           //PostuladosN = db.Postulaciones.Where(p => p.RequisicionId.Equals(e.Id) & p.StatusId.Equals(1)).Select(p => new
-                           //{
-                           //    p.CandidatoId,
-                           //    p.Candidato.Nombre,
-                           //    p.Candidato.ApellidoPaterno,
-                           //    p.Candidato.ApellidoMaterno,
-                           //    p.Candidato.CURP,
-                           //    email = p.Candidato.emails.Select(m => m.email).FirstOrDefault(),
-                           //    p.StatusId
-                           //}),
-                           //EnProceso = db.ProcesoCandidatos.Where(p => p.RequisicionId.Equals(e.Id) && p.EstatusId != 27 && p.EstatusId != 40 && p.EstatusId != 28 && p.EstatusId != 42).Count(),
-                           //EnProcesoN = db.ProcesoCandidatos.Where(p => p.RequisicionId.Equals(e.Id) && p.EstatusId != 24 && p.EstatusId != 27 && p.EstatusId != 40 && p.EstatusId != 28 && p.EstatusId != 42).Select(d => new
-                           //{
-                           //    candidatoId = d.CandidatoId,
-                           //    nombre = db.Candidatos.Where(x => x.Id.Equals(d.CandidatoId)).Select(cc => cc.Nombre + " " + cc.ApellidoPaterno + " " + cc.ApellidoMaterno).FirstOrDefault(),
-                           //    email = db.Emails.Where(x => x.EntidadId.Equals(d.CandidatoId)).Select(m => m.email).FirstOrDefault(),
-                           //    estatusId = d.EstatusId
-                           //}),
-                           //Contratados = db.ProcesoCandidatos.Where(p => p.RequisicionId.Equals(e.Id) && p.EstatusId.Equals(24)).Count(),
-                           Propietario = db.Usuarios.Where(x => x.Id.Equals(e.PropietarioId)).Select(P => P.Nombre + " " + P.ApellidoPaterno + " " + P.ApellidoMaterno).FirstOrDefault(),
-                           reclutadores = db.AsignacionRequis.Where(x => x.RequisicionId.Equals(e.Id) && !x.GrpUsrId.Equals(e.AprobadorId)).Select(a => new
-                           {
-                               reclutador = db.Usuarios.Where(x => x.Id.Equals(a.GrpUsrId)).Select(r => r.Nombre + " " + r.ApellidoPaterno + " " + r.ApellidoMaterno).FirstOrDefault()
-                           }).Distinct().ToList(),
-                           //ComentarioReclutador = db.ComentariosVacantes.Where(x => x.RequisicionId.Equals(e.Id)).Select(c => c.fch_Creacion + " - " + c.UsuarioAlta + " - " + (c.Motivo.Id == 7 ? "" : c.Motivo.Descripcion + " - ") + c.Comentario).ToList()
-                       }).OrderBy(x => x.EstatusOrden).ThenByDescending(x => x.Folio).ToList();
-
-                    return Ok(requisicionMTY);
-                }
+                
 
             }
             catch (Exception ex)
@@ -1982,8 +1885,8 @@ namespace SAGA.API.Controllers
                 requisicion.fch_Modificacion = DateTime.Now;
                 if (requisicion.Publicado)
                     requisicion.Publicado = false;
-
-                db.AsignacionRequis.RemoveRange(asignados);
+                if(asignados.Count > 0)
+                    db.AsignacionRequis.RemoveRange(asignados);
 
 
 
@@ -1993,7 +1896,7 @@ namespace SAGA.API.Controllers
                 Guid trazabilidadId = db.TrazabilidadesMes.Where(x => x.Folio.Equals(Folio)).Select(x => x.Id).FirstOrDefault();
                 //Isertar el registro de la rastreabilidad. 
                 rastreabilidad.RastreabilidadInsert(trazabilidadId, requi.UsuarioMod, 6);
-
+                
                 SendEmail.ConstructEmail(asignados, null, "RU", Folio, string.Empty, VBra, null);
 
                 db.SaveChanges();
